moved mock data and models to their respective files
