+++ src/app/core/mock-data/expense.data.ts
@@ -5524,3 +5524,77 @@ export const newExpenseViewReport: Expense[] = [
   },
   { ...expenseData3, tx_report_id: 'rpkkdCNBFsJ2' },
 ];
+
+export const expectedFormattedTransaction: Partial<Expense>[] = [
+  {
+    tx_created_at: new Date('2023-02-08T06:47:48.414Z'),
+    tx_updated_at: new Date('2023-02-08T06:47:48.414Z'),
+    tx_id: 'txNVtsqF8Siq',
+    tx_source: 'MOBILE_DASHCAM_BULK',
+    tx_org_user_id: 'ouX8dwsbLCLv',
+    tx_creator_id: 'ouX8dwsbLCLv',
+    tx_txn_dt: new Date('2023-02-08T17:00:00.000Z'),
+    tx_vendor: null,
+    tx_platform_vendor: null,
+    tx_category: null,
+    tx_amount: null,
+    tx_currency: null,
+    tx_admin_amount: null,
+    tx_user_amount: null,
+    tx_orig_amount: null,
+    tx_orig_currency: null,
+    tx_tax: null,
+    tx_tax_amount: null,
+    tx_tax_group_id: null,
+    tx_report_id: null,
+    tx_reported_at: null,
+    tx_num_files: 0,
+    tx_invoice_number: null,
+    tx_purpose: null,
+    tx_vendor_id: null,
+    tx_platform_vendor_id: null,
+    tx_project_id: null,
+    tx_billable: null,
+    tx_skip_reimbursement: false,
+    tx_state: 'DRAFT',
+    tx_external_id: null,
+    tx_cost_center_id: null,
+    tx_payment_id: null,
+    tx_source_account_id: 'acc5APeygFjRd',
+    tx_org_category_id: 16582,
+    tx_physical_bill: null,
+    tx_policy_state: null,
+    tx_manual_flag: null,
+    tx_policy_flag: null,
+    tx_expense_number: 'E/2023/02/T/51',
+    tx_split_group_id: 'txNVtsqF8Siq',
+    tx_split_group_user_amount: null,
+    tx_extracted_data: null,
+    tx_mandatory_fields_present: false,
+    tx_user_reason_for_duplicate_expenses: null,
+    tx_distance: null,
+    tx_distance_unit: null,
+    tx_from_dt: null,
+    tx_to_dt: null,
+    tx_num_days: null,
+    tx_fyle_category: 'Unspecified',
+    tx_mileage_calculated_distance: null,
+    tx_mileage_calculated_amount: null,
+    tx_mileage_vehicle_type: null,
+    tx_mileage_rate: null,
+    tx_mileage_is_round_trip: null,
+    tx_hotel_is_breakfast_provided: null,
+    tx_flight_journey_travel_class: null,
+    tx_flight_return_travel_class: null,
+    tx_train_travel_class: null,
+    tx_bus_travel_class: null,
+    tx_taxi_travel_class: null,
+    tx_per_diem_rate_id: null,
+    tx_activity_policy_pending: null,
+    tx_activity_details: null,
+    tx_custom_properties: null,
+    tx_physical_bill_at: null,
+    tx_policy_amount: null,
+    tx_locations: [],
+  },
+];
