+++ src/app/core/mock-data/unformatted-transaction.data.ts
@@ -1,0 +1,75 @@
+import { UnformattedTransaction } from '../models/unformatted-transaction.model';
+
+export const unformattedTxnData: UnformattedTransaction[] = [
+  {
+    created_at: new Date('2023-02-08T06:47:48.414Z'),
+    updated_at: new Date('2023-02-08T06:47:48.414Z'),
+    id: 'txNVtsqF8Siq',
+    source: 'MOBILE_DASHCAM_BULK',
+    org_user_id: 'ouX8dwsbLCLv',
+    creator_id: 'ouX8dwsbLCLv',
+    txn_dt: new Date('2023-02-08T17:00:00.000Z'),
+    vendor: null,
+    platform_vendor: null,
+    category: null,
+    amount: null,
+    currency: null,
+    admin_amount: null,
+    user_amount: null,
+    orig_amount: null,
+    orig_currency: null,
+    tax: null,
+    tax_amount: null,
+    tax_group_id: null,
+    report_id: null,
+    reported_at: null,
+    num_files: 0,
+    invoice_number: null,
+    purpose: null,
+    vendor_id: null,
+    platform_vendor_id: null,
+    project_id: null,
+    billable: null,
+    skip_reimbursement: false,
+    state: 'DRAFT',
+    external_id: null,
+    cost_center_id: null,
+    payment_id: null,
+    source_account_id: 'acc5APeygFjRd',
+    org_category_id: 16582,
+    physical_bill: null,
+    policy_state: null,
+    manual_flag: null,
+    policy_flag: null,
+    expense_number: 'E/2023/02/T/51',
+    split_group_id: 'txNVtsqF8Siq',
+    split_group_user_amount: null,
+    extracted_data: null,
+    mandatory_fields_present: false,
+    user_reason_for_duplicate_expenses: null,
+    distance: null,
+    distance_unit: null,
+    from_dt: null,
+    to_dt: null,
+    num_days: null,
+    fyle_category: 'Unspecified',
+    mileage_calculated_distance: null,
+    mileage_calculated_amount: null,
+    mileage_vehicle_type: null,
+    mileage_rate: null,
+    mileage_is_round_trip: null,
+    hotel_is_breakfast_provided: null,
+    flight_journey_travel_class: null,
+    flight_return_travel_class: null,
+    train_travel_class: null,
+    bus_travel_class: null,
+    taxi_travel_class: null,
+    per_diem_rate_id: null,
+    activity_policy_pending: null,
+    activity_details: null,
+    custom_properties: null,
+    physical_bill_at: null,
+    policy_amount: null,
+    locations: [],
+  },
+];
