+++ src/app/core/models/card-aggregate-stat.model.ts
@@ -1,4 +1,4 @@
-export interface CardAggregateStat {
+export interface CardAggregateStats {
   aggregates: { function_name: string; function_value: number }[];
   key: { column_name: string; column_value: string }[];
 }
